[artf155298] - Updates on deletion of the config, minor refactoring, logging

+++ simple2secure.portal/src/main/java/com/simple2secure/portal/utils/GroupUtils.java
@@ -16,6 +16,9 @@ import com.simple2secure.api.model.CompanyLicensePrivate;
 import com.simple2secure.api.model.Context;
 import com.simple2secure.api.model.ContextUserAuthentication;
 import com.simple2secure.api.model.GroupAccessRight;
+import com.simple2secure.api.model.Processor;
+import com.simple2secure.api.model.QueryRun;
+import com.simple2secure.api.model.Step;
 import com.simple2secure.api.model.User;
 import com.simple2secure.api.model.UserRole;
 import com.simple2secure.commons.config.StaticConfigItems;
@@ -129,36 +132,40 @@ public class GroupUtils {
 	 *
 	 * @param groupId
 	 */
-	public void deleteGroup(String groupId) {
-		// Delete the group configurations
-		configRepository.deleteByGroupId(groupId);
-
+	public void deleteGroup(String groupId, boolean deleteAll) {
 		// Delete the group steps
 		stepRepository.deleteByGroupId(groupId);
 
 		// Delete the group processors
 		processorRepository.deleteByGroupId(groupId);
 
-		// Delete all licenses and all probe reports which were created
-		List<CompanyLicensePrivate> licenses = licenseRepository.findByGroupId(groupId);
-
-		if (licenses != null) {
-			for (CompanyLicensePrivate license : licenses) {
-				if (!Strings.isNullOrEmpty(license.getProbeId())) {
-					reportRepository.deleteByProbeId(license.getProbeId());
-					networkReportRepository.deleteByProbeId(license.getProbeId());
-				}
-				licenseRepository.delete(license);
-			}
-		}
-
 		// Remove OSQuery configuration
 		queryRepository.deleteByGroupId(groupId);
 
-		// Remove GroupAccessRights
-		groupAccessRightRepository.deleteByGroupId(groupId);
+		// if this flag is set all group dependencies will be deleted
+		if (deleteAll) {
+			// Delete all licenses and all probe reports which were created
+			List<CompanyLicensePrivate> licenses = licenseRepository.findByGroupId(groupId);
 
-		deleteGroupFromChildren(groupId);
+			if (licenses != null) {
+				for (CompanyLicensePrivate license : licenses) {
+					if (!Strings.isNullOrEmpty(license.getProbeId())) {
+						reportRepository.deleteByProbeId(license.getProbeId());
+						networkReportRepository.deleteByProbeId(license.getProbeId());
+					}
+					licenseRepository.delete(license);
+				}
+			}
+
+			// Remove GroupAccessRights
+			groupAccessRightRepository.deleteByGroupId(groupId);
+
+			// Delete the group configurations
+			configRepository.deleteByGroupId(groupId);
+
+			deleteGroupFromChildren(groupId);
+		}
+
 	}
 
 	/**
@@ -171,7 +178,7 @@ public class GroupUtils {
 			List<CompanyGroup> groups = groupRepository.findByContextId(contextId);
 			if (groups != null) {
 				for (CompanyGroup group : groups) {
-					deleteGroup(group.getId());
+					deleteGroup(group.getId(), true);
 				}
 			}
 		}
@@ -226,9 +233,9 @@ public class GroupUtils {
 		for (CompanyGroup group : children) {
 			if (PortalUtils.groupHasChildren(group)) {
 				deleteGroupChildren(group);
-				deleteGroup(group.getId());
+				deleteGroup(group.getId(), true);
 			} else {
-				deleteGroup(group.getId());
+				deleteGroup(group.getId(), true);
 			}
 		}
 	}
@@ -463,6 +470,42 @@ public class GroupUtils {
 		}
 
 		return groupIds;
+	}
+
+	/**
+	 * This function copies the configuration from the source group to the destination group
+	 * 
+	 * @param sourceGroupId
+	 * @param destGroupId
+	 */
+	public void copyGroupConfiguration(String sourceGroupId, String destGroupId) {
+		List<QueryRun> queries = queryRepository.findByGroupId(sourceGroupId, true);
+		List<Processor> processors = processorRepository.getProcessorsByGroupId(sourceGroupId);
+		List<Step> steps = stepRepository.getStepsByGroupId(sourceGroupId, true);
+
+		if (queries != null) {
+			for (QueryRun query : queries) {
+				query.setGroupId(destGroupId);
+				query.setId(null);
+				queryRepository.save(query);
+			}
+		}
+
+		if (processors != null) {
+			for (Processor processor : processors) {
+				processor.setGroupId(destGroupId);
+				processor.setId(null);
+				processorRepository.save(processor);
+			}
+		}
+
+		if (steps != null) {
+			for (Step step : steps) {
+				step.setGroupId(destGroupId);
+				step.setId(null);
+				stepRepository.save(step);
+			}
+		}
 	}
 
 	/**
